Fix lints and tests

+++ x-pack/libbeat/reader/etw/provider_test.go
@@ -71,7 +71,7 @@ func TestGUIDFromProviderName_GUIDNotFound(t *testing.T) {
 		// Calculate size needed for the provider name string
 		nameSize := (len(utf16ProviderName) + 1) * 2 // +1 for null-terminator
 
-		requiredSize := uint32(unsafe.Sizeof(ProviderEnumerationInfo{})) + uint32(unsafe.Sizeof(TraceProviderInfo{})) + uint32(nameSize)
+		requiredSize := uint32(unsafe.Sizeof(ProviderEnumerationInfo{}) + unsafe.Sizeof(TraceProviderInfo{}) + uintptr(nameSize))
 		if *pBufferSize < requiredSize {
 			*pBufferSize = requiredSize
 			return ERROR_INSUFFICIENT_BUFFER
@@ -123,7 +123,7 @@ func TestGUIDFromProviderName_Success(t *testing.T) {
 		// Calculate size needed for the provider name string
 		nameSize := (len(utf16ProviderName) + 1) * 2 // +1 for null-terminator
 
-		requiredSize := uint32(unsafe.Sizeof(ProviderEnumerationInfo{})) + uint32(unsafe.Sizeof(TraceProviderInfo{})) + uint32(nameSize)
+		requiredSize := uint32(unsafe.Sizeof(ProviderEnumerationInfo{}) + unsafe.Sizeof(TraceProviderInfo{}) + uintptr(nameSize))
 		if *pBufferSize < requiredSize {
 			*pBufferSize = requiredSize
 			return ERROR_INSUFFICIENT_BUFFER
